backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
transformers==4.36.0
torch==2.2.0
sentencepiece==0.2.0
pydantic==2.5.0
requests==2.31.0